sane versioning standards a la http://semver.org, v0.10.0-rc2

+++ version_update.sh
@@ -5,25 +5,23 @@
 #
 DATE_STR=`date +"%a, %d %b %Y %H:%M:%S +0100"`
 VERSION_FILES=('setup.py' 'source/RPIO/__init__.py' 'source/c_gpio/py_gpio.c' 'source/c_pwm/pwm_py.c' 'documentation/source/conf.py');
-VERSION_LAST=`head -n1 debian/changelog | awk '{print $2}' | sed "s/[()]//g"`
+VERSION_LAST=`cat VERSION`
 
 if [ "$1" == "--show" ]; then
     echo $VERSION_LAST
     exit 0
 fi
 
-echo "The last version is $VERSION_LAST."
-echo -n "New version number: "
+echo "Current Version: $VERSION_LAST"
+echo -n "    New version: "
 read version
-
 if [ -e $version ]; then
     exit 1
 fi
 
-echo -n "Are you sure you want to update the project to v$version? [y/n] "
+echo -n "Are you sure you want to update the project? [y/n] "
 read confirm
-
-if [ -e $confirm ]; then
+if [ "$confirm" != "y" ]; then
     exit 1
 fi
 
@@ -34,18 +32,24 @@ for fn in ${VERSION_FILES[@]}; do
     sed -i '' "s/$VERSION_LAST/$version/g" $fn
 done
 
-# Update Changelog
-echo "rpio ($version) unstable; urgency=low" > CHANGELOG.new
-echo "" >> CHANGELOG.new
-echo "  * " >> CHANGELOG.new
-echo "  * " >> CHANGELOG.new
-echo "  * " >> CHANGELOG.new
-echo "" >> CHANGELOG.new
-echo " -- Chris Hager <chris@linuxuser.at>  $DATE_STR" >> CHANGELOG.new
-echo "" >> CHANGELOG.new
-cat debian/changelog >> CHANGELOG.new
-mv CHANGELOG.new debian/changelog
-nano -w debian/changelog
+echo $version > VERSION
+
+echo -n "Do you want to update debian/changelog? [y/n] "
+read confirm
+if [ "$confirm" == "y" ]; then
+    # Update Changelog
+    echo "rpio ($version) unstable; urgency=low" > CHANGELOG.new
+    echo "" >> CHANGELOG.new
+    echo "  * " >> CHANGELOG.new
+    echo "  * " >> CHANGELOG.new
+    echo "  * " >> CHANGELOG.new
+    echo "" >> CHANGELOG.new
+    echo " -- Chris Hager <chris@linuxuser.at>  $DATE_STR" >> CHANGELOG.new
+    echo "" >> CHANGELOG.new
+    cat debian/changelog >> CHANGELOG.new
+    mv CHANGELOG.new debian/changelog
+    nano -w debian/changelog
+fi
 
 # Now rebuild doc for latest version
 echo "Run 'make doc' now, to build documentation with the new version number"
